// web/src/gen/model/tagLinkDetailDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface TagLinkDetailDto {
  tagLinkId: string;
  tagId: string;
  tagPath: string;
  tagName: string;
  /** @nullable */
  tagColor: string | null;
  /** @nullable */
  tagDescription: string | null;
  /** @nullable */
  targetSubPath: string | null;
  /** @nullable */
  metadataJson: string | null;
}